--- src/pages/Cart/Button.scss
@@ -1,39 +1,0 @@
-.button-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 40px;
-  height: 15px;
-
-  margin: 10px;
-
-  .count {
-    input {
-      width: 24px;
-      height: 22.7px;
-      border: 1px solid #d4d8d9;
-      border-radius: 3px 0 0 3px;
-      padding-left: 30%;
-      &:focus {
-        outline: none;
-      }
-    }
-  }
-  .count-btn {
-    display: flex;
-    flex-direction: column;
-    button {
-      width: 17px;
-      height: 11.8px;
-      border: none;
-      background-color: #fff;
-      padding: 0;
-      cursor: pointer;
-      img {
-        width: 100%;
-        height: 100%;
-      }
-    }
-  }
-}
